Stop Passenger preloader from idle
When combined with a non-zero MaxRequests setting this results
in the preloader potentially shutting down / idling, and semi-frequent
forks happening slowly that can cause 503 errors that we've been
seeing.

Holding off on changing MaxRequests to zero as well since we are at
a big EOQ during midterms, but after that the recommendation is to
set MaxRequests to zero as well, plus overall tune Passenger and
upgrade it to a newer version.

Hoping to observe 503s drastically reduced in production after this
is converged.

There is a Google Docs presentation by David Fisher covering the process
at: https://docs.google.com/presentation/d/1CSn5IzghkRBAc1Rfx5yFI3r-9_AajE9YD6d05oMcWKQ